Add contour.py and rune-detector.py with contour detection algorithm.
Update contour detecting algorithm by using BGR2HSV instead of
BGR2Grayscale. It achieves great performance in detecting red color.

Apply this to detect the rune in the video in rune-detector.py. But this
algorithm vastly depends on the color range and contour area, which are
parameters need to tune every time in a different scenario.