Add 2 projects python script.
Project 1 is the color detection. Use HSV color space with the webcam to
detect yellow objects.

Project 2 is the face anonymizer. Use mediapipe library embed face
detection algorithm to detect faces in the image and then blur it by
replaing the nparray elements. It works well on webcam and the video can
be saved locally. HEAD UP: when detect face on the iPhone screen,
unexpected errors will exist, which is to be solved.